Posts refactoring. Unite feed posts and profile posts in united component
