+++ src/components/Feed/Posts/Posts.tsx
@@ -1,28 +1,23 @@
 import { FC, useState } from 'react';
 import AddNewPost from './AddNewPost/AddNewPost';
-import { FeedPostItemType } from "../../../types/types";
-import OpenPostBlock from './PostItem/OpenPost/OpenPostBlock';
-import PostItem from './PostItem/PostItem';
+import { FeedPostItemType, postItemType } from "../../../types/types";
+import PostItem from '../../PostItem/PostItem';
 
 type Props = {
     FeedPosts: any
 }
 
 const Posts: FC<Props> = (props) => {
-    debugger;
-    const [openPost, setOpenPost] = useState<boolean>(false);
-    return (<div>
-        <div>
-            <AddNewPost />
-        </div>
-        {openPost && <OpenPostBlock setOpenPost={(openMode: boolean) => setOpenPost(openMode)}/>}
-        <div className="flex flex-col items-center">
-            {props.FeedPosts.map((post: FeedPostItemType) => {
-                return <PostItem post={post} setOpenPost={(openPost: boolean) => {setOpenPost(openPost)}}/>
-            })
-            }
-        </div>
-    </div>)
-}
+  return (
+    <div>
+      <div>
+        <AddNewPost />
+      </div>
+      <div className="flex flex-col items-center">
+        {props.FeedPosts.map((post: postItemType) => <PostItem post={post} />)}
+      </div>
+    </div>
+  );
+};
 
 export default Posts;
