+++ src/components/PostItem/PostInterectionsBlock/Interections.css
