+++ src/components/Profile/PostsBlock/Posts.tsx
@@ -1,8 +1,8 @@
 import { FC, memo, useState } from 'react';
-import PostItem from './Post/PostItem';
 import AddingNewPost from './Post/AddingNewPost/AddingNewPost';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../../redux/store';
+import PostItem from '../../PostItem/PostItem';
 
 const Posts: FC = () => {
     const postItem = useSelector((state: RootState) => state.ProfilePage.PostItem);
